[FEAT] 원본 레스토랑 엔티티 정의
