+++ src/main/java/com/lucky/around/meal/datapipeline/RawRestaurant.java
@@ -1,0 +1,30 @@
+package com.lucky.around.meal.datapipeline;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+@Builder
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class RawRestaurant {
+
+  @Id private String id;
+
+  private String jsonData;
+
+  private String hash;
+
+  private boolean isUpdated;
+
+  @Builder
+  public RawRestaurant(String id, String jsonData, String hash, boolean isUpdated) {
+    this.id = id;
+    this.jsonData = jsonData;
+    this.hash = hash;
+    this.isUpdated = isUpdated;
+  }
+}
